Phase 2A
Each assembled source file has a corresponding version file which is "bumped" with each change.

The resulting "*.a.i" file can be included to embed the file version infomration in the output binary.

Although this functionality was rarely used, an example can be seen here (14 seconds into the video, the version of "xb" is displayed):

https://www.youtube.com/watch?v=-P3JGxBNUyM

The pair of scripts "rlvers" and "rlvers0" generate a listing file "rl.ver" which indicates the version of each source file.